+++ linkis-public-enhancements/linkis-publicservice/linkis-udf/src/main/java/com/webank/wedatasphere/linkis/udf/dao/UDFTreeDao.java
@@ -1,0 +1,36 @@
+/*
+ * Copyright 2019 WeBank
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.webank.wedatasphere.linkis.udf.dao;
+
+import com.webank.wedatasphere.linkis.udf.entity.UDFTree;
+
+import java.util.List;
+import java.util.Map;
+
+
+public interface UDFTreeDao {
+
+    void addTree(UDFTree udfTree);
+
+    void updateTree(UDFTree udfTree);
+
+    void deleteTree(Long id,  String userName);
+
+    UDFTree getTreeById(Long id, String category);
+
+    List<UDFTree> getTreesByParentId(Map<String, Object> params);
+}
