+++ linkis-public-enhancements/linkis-publicservice/linkis-udf/src/main/java/com/webank/wedatasphere/linkis/udf/utils/ConstantVar.java
@@ -1,0 +1,30 @@
+/*
+ * Copyright 2019 WeBank
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.webank.wedatasphere.linkis.udf.utils;
+
+public class ConstantVar {
+
+    public final static int UDF_JAR = 0;
+    public final static int UDF_PY = 1;
+    public final static int UDF_SCALA = 2;
+    public final static int FUNCTION_PY = 3;
+    public final static int FUNCTION_SCALA = 4;
+
+    public final static String FUNCTION = "function";
+    public final static String UDF = "udf";
+    public final static String ALL = "all";
+}
